Update __isset to check for key on Presenter
Using functions like "lists" on Laravel Collections checks isset to
determine whether an object has a particular key. It is useful (for
example in listboxes) to be able to specify a key that is not present on
the underlying model, as it is purely for presentation purposes.

Currently this is not possible, as the __isset call will return false
unless the specified key exists on the wrappedObject, even though a
corresponding method may exist on the Presenter.

This change adds a check for the method name on the Presenter, before
checking the wrappedObject